Constants Update and Created Lift
Fixed spelling errors and added variables for an upper and lower limit in Constants.

Created new instances of motors for wrist and elevator. Added an open and closed loop function for both the wrist and elevator. As well as the position and error for both the wrist and elevator.
Also created a SmartDashboard instance for each corresponding position and error.